Extract endpoint specific code
