+++ corehq/apps/case_search/utils.py
@@ -2,7 +2,6 @@ import json
 import re
 from collections import defaultdict
 from dataclasses import dataclass
-from datetime import date
 from functools import wraps
 
 from django.conf import settings
@@ -17,8 +16,6 @@ from casexml.apps.phone.data_providers.case.livequery import (
     get_all_related_live_cases,
 )
 from dimagi.utils.logging import notify_exception
-
-from sqlalchemy import and_, cast, func, literal, not_, or_, select
 
 from corehq import toggles
 from corehq.apps.app_manager.dbaccessors import get_app_cached
@@ -53,7 +50,9 @@ from corehq.apps.case_search.models import (
     CaseSearchEndpoint,
     extract_search_request_config,
 )
-from corehq.apps.case_search.endpoint_query_spec import ParameterInput, parse_parameter_spec, parse_query_spec
+from corehq.apps.case_search.endpoint_query_spec import parse_parameter_spec, parse_query_spec
+from corehq.apps.case_search.endpoint_sql_query_builder import get_sql_endpoint_results
+from corehq.apps.case_search.query_builder_base import BaseCaseSearchEndpointQueryBuilder
 from corehq.apps.case_search.xpath_functions.query_functions import date_permutations, validate_date
 from corehq.apps.es import HQESQuery, case_search
 from corehq.apps.es import cases as case_es
@@ -71,10 +70,6 @@ from corehq.apps.es.case_search import (
     wrap_case_search_hit,
 )
 from corehq.apps.es.profiling import ESQueryProfiler
-
-from corehq.apps.project_db.populate import coerce_to_date, coerce_to_gps, coerce_to_number, coerce_to_select
-from corehq.apps.project_db.query import rows_to_cases, to_distance_in_meters
-from corehq.apps.project_db.table_ddl import CaseTable, Earth, get_project_db_engine, property_column
 
 from corehq.apps.registry.exceptions import (
     RegistryAccessException,
@@ -179,21 +174,7 @@ def get_primary_case_search_endpoint_results(helper, target_type,
         with helper.profiler.timing_context('wrap_cases'):
             cases = [helper.wrap_case(hit, include_score=True) for hit in results.raw_hits]
     else:
-        builder = CaseSearchEndpointSqlQueryBuilder(helper, case_type, endpoint_query)
-        with helper.profiler.timing_context('build_query'):
-            query = builder.build_query(criteria)
-        if limit:
-            query = query.limit(limit)
-
-        engine = get_project_db_engine()
-
-        with engine.begin() as conn:
-            # In sqlalchemy 1.4+, use execution_options postgresql_readonly
-            # conn.execute(sqlalchemy.text('SET TRANSACTION READ ONLY'))
-            # DomainSchema(domain).set_local_search_path(conn)
-            result = conn.execute(query)
-            rows = result.fetchall()
-            cases = rows_to_cases(rows, builder.table, builder.request_domain, case_type)
+        cases = get_sql_endpoint_results(helper, case_type, criteria, endpoint_query, limit)
 
     helper.profiler.primary_count = len(cases)
     return cases
@@ -431,61 +412,6 @@ class CaseSearchQueryBuilder:
             for prop in properties_config.properties
         ]
 
-class BaseCaseSearchEndpointQueryBuilder:
-    """Walks a parsed query-builder AST (GroupNode/ComponentNode) and
-    resolves parameter references, deferring backend-specific leaf
-    translation and boolean combination to subclasses (e.g. Elasticsearch,
-    SQL).
-    """
-    def __init__(self, query_root):
-        self.query_root = query_root
-
-    def _parse_query_root(self, search_criteria):
-        self.param_values = {c.key: c.value for c in search_criteria}
-        return self._parse_query(self.query_root)
-
-    def _get_child_queries(self, node):
-        child_queries = [self._parse_query(child) for child in node.children]
-        return [q for q in child_queries if q is not None]
-
-    def _parse_query(self, node):
-        if node.type in ('all', 'any', 'none'):
-            # Drop a group with no surviving children rather than collapsing to
-            # an empty AND/OR. An empty bool matches all documents, which in an
-            # `any`/OR context would make the whole query match everything.
-            children = self._get_child_queries(node)
-            if not children:
-                return None
-            if node.type == 'all':
-                return self._combine_and(children)
-            if node.type == 'any':
-                return self._combine_or(children)
-            return self._combine_none(children)
-        elif node.type == 'component':
-            return self._parse_component_node(node)
-        else:
-            return None
-
-    def _input_value(self, input_):
-        if input_ is None:
-            return None
-        if isinstance(input_, ParameterInput):
-            return self.param_values.get(input_.value)
-        return input_.value
-
-    def _combine_and(self, children):
-        raise NotImplementedError
-
-    def _combine_or(self, children):
-        raise NotImplementedError
-
-    def _combine_none(self, children):
-        raise NotImplementedError
-
-    def _parse_component_node(self, node):
-        raise NotImplementedError
-
-
 class CaseSearchEndpointQueryBuilder(BaseCaseSearchEndpointQueryBuilder):
     """Compiles the case search object for the view"""
     def __init__(
@@ -587,123 +513,6 @@ class CaseSearchEndpointQueryBuilder(BaseCaseSearchEndpointQueryBuilder):
             elif operator == 'phonetic':
                 return sounds_like_text_query(field, value)
         return None
-
-class CaseSearchEndpointSqlQueryBuilder(BaseCaseSearchEndpointQueryBuilder):
-    def __init__(
-        self,
-        helper,
-        case_type,
-        query_root):
-        super().__init__(query_root)
-        self.request_domain = helper.domain
-        self.case_type = case_type
-        self.helper = helper
-        self.config = helper.config
-        self.table = CaseTable(self.request_domain, self.case_type).reflect()
-
-    def build_query(self, search_criteria):
-        query = self._get_initial_query()
-        where_clause = self._parse_query_root(search_criteria)
-        if where_clause is None:
-            # Every condition dropped (e.g. all inputs were unsupplied
-            # parameters). Apply no extra filter rather than match-all-via-empty.
-            return query
-        return query.where(where_clause)
-
-    def _get_initial_query(self):
-        max_results = CASE_SEARCH_MAX_RESULTS
-        if toggles.INCREASED_MAX_SEARCH_RESULTS.enabled(self.request_domain):
-            max_results = 1500
-
-        return select(self.table.columns).limit(max_results)
-
-    def _combine_and(self, children):
-        return and_(*children)
-
-    def _combine_or(self, children):
-        return or_(*children)
-
-    def _combine_none(self, children):
-        return not_(and_(*children))
-
-    def _parse_component_node(self, node):
-        operator = node.operator
-        column = self.table.columns[property_column(node.field, node.field_type)]
-        value = self._input_value(node.inputs['value'])
-        if value is None:
-            return None  # ignore component if value is not given
-
-        if node.field_type == FIELD_TYPE_GPS:
-            if operator == 'within_distance':
-                point = coerce_to_gps(self._input_value(node.inputs.get('point')))
-                distance = self._input_value(node.inputs.get('distance'))
-                unit = self._input_value(node.inputs.get('unit'))
-                print(f"point: {point}, distance: {distance}, unit: {unit}")
-                if None in (point, distance, unit):
-                    return None
-                if unit not in queries.DISTANCE_UNITS:
-                    return None
-                earth_point = cast(literal(point), Earth)
-                distance = to_distance_in_meters(distance, unit)
-                return func.earth_distance(column, earth_point) <= distance
-            return None
-
-        elif node.field_type in (FIELD_TYPE_DATE, FIELD_TYPE_DATETIME):
-            date_value = coerce_to_date(value)
-            if date_value is None:
-                return None
-            if operator == 'equals':
-                return column == date_value
-            elif operator == 'lt':
-                return column < date_value
-            elif operator == 'gt':
-                return column > date_value
-            elif operator == 'lte':
-                return column <= date_value
-            elif operator == 'gte':
-                return column >= date_value
-            elif operator == 'fuzzy_date':
-                return column.in_([date.fromisoformat(p) for p in date_permutations(value)])
-            return None
-        elif node.field_type == FIELD_TYPE_NUMBER:
-            number_value = coerce_to_number(value)
-            if number_value is None:
-                return None
-            if operator == 'equals':
-                return column == value
-            elif operator == 'not_equals':
-                return column != value
-            elif operator == 'lt':
-                return column < value
-            elif operator == 'gt':
-                return column > value
-            elif operator == 'lte':
-                return column <= value
-            elif operator == 'gte':
-                return column >= value
-            return None
-        elif node.field_type == FIELD_TYPE_SELECT:
-            arr_value = coerce_to_select(value)
-            print(arr_value)
-            if operator == 'selected_any':
-                return column.overlap(arr_value)
-            elif operator == 'selected_all':
-                return column.contains(arr_value)
-            elif operator == 'is_empty':
-                return func.cardinality(column) == 0
-            return None
-        else:
-            if operator == 'equals':
-                return column == value
-            elif operator == 'not_equals':
-                return column != value
-            elif operator == 'starts_with':
-                return column.startswith(value)
-            elif operator == 'fuzzy':
-                return func.similarity(column, value) >= 0.3
-            elif operator == 'phonetic':
-                return func.dmetaphone(column) == func.dmetaphone(value)
-            return None
 
 
 @time_function()
